+++ crates/core/src/commands/pty/mod.rs
@@ -116,7 +116,7 @@ pub async fn interactive_command_with_env<S: AsRef<str> + Sync>(
         "{starting}{command} {flags} {IO_SUBS} && {ending}",
         command = arguments.command_string.as_ref(),
         flags = match arguments.output_mode {
-            ChildOutputMode::Nix => "--log-format internal-json",
+            ChildOutputMode::Nix(..) => "--log-format internal-json",
             ChildOutputMode::Generic => "",
         },
         starting = create_starting_segment(&needles.start),
@@ -162,7 +162,7 @@ pub async fn interactive_command_with_env<S: AsRef<str> + Sync>(
             began_tx,
             reader,
             needles,
-            output_mode: arguments.output_mode,
+            output_mode: arguments.output_mode.clone(),
             stderr_collection: stderr_collection.clone(),
             stdout_collection: stdout_collection.clone(),
             span: Span::current(),