+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - yyyy-mm-dd
 
+### Added
+
+- The last associated nix log is now shown next to the node name, including a
+  possible build name when `--print-build-logs` is passed.
 - Fix outputting coloured text, disrespecting NO_COLOR, in some case.
 
 ### Changed